config refactoring
Finally fixed the build pipeline so the stand alone UI mod can be generated. Simplified the config by removing the option to build the gameplay mod using the UI mod as a base, because it was just complicating things to save a trivial amount of build time.

Improved docstrings of all editor functions to show the whole filepath of the file being edited.

Removed redundant "ui_dev" and "ui_release" config parameters, not sure how that happened.

TODO:
1. Add new divs to the hide_divs config parameter
2. Figure out how to handle functions like edit_fob_attributes, which have a mix of gameplay and ui edits. Can either separate them into two functions or maybe pass build_target_cfg from main() to the respective function.