v22.5 (hotfix)
Fixed crash caused by Chasseurs_CMD2_FR. Depiction code needs improvement still.

Updated dictionaries with a new "UnitRole" key. The value is just whatever the first value in the SpecialtiesList was previously. This was something Eugen changed, and it dictates what the main icon is for the unit.

strategicdecks.py (new) - Currently only used to update the Army General decks to use the new command units instead of the leader units. Army General should be playable now in the mod.