Infantry armor fix
I forgot to finish implementing the new method to apply the custom armor for small arms to infantry units. The old script used a bunch of conditional logical based on the type of unit, but there was always uncaught special cases which I had to add exceptions for. So I changed it to just use dictionary references and manually defined each infantry unit to have its armor edited to that reference value.